Obsolescence and formatting improved for version 1.4.0. (Pre-preparation for 2.0)
