+++ ChustaSoft.Common/Contracts/ITrackableBuilder.cs
@@ -7,7 +7,7 @@ namespace ChustaSoft.Common.Contracts
         : IBuilder<T> where T : class
     {
 
-        //ICollection<ErrorMessage> Errors { get; }
+        //ICollection<ErrorMessage> Errors { get; } //TODO: Pending to uncomment in version 2.0 once the IBuilder property for Errors has been removed
 
     }
 }
