+++ ChustaSoft.Common/Helpers/CollectionsHelper.cs
@@ -5,7 +5,6 @@ using System.Collections.Concurrent;
 using System.Collections.Generic;
 using System.Linq;
 
-
 namespace ChustaSoft.Common.Helpers
 {
     /// <summary>
