+++ ChustaSoft.Common/Helpers/EnumsHelper.cs
@@ -4,7 +4,6 @@ using System.Collections.Generic;
 using System.ComponentModel;
 using System.Linq;
 
-
 namespace ChustaSoft.Common.Helpers
 {
