+++ ChustaSoft.Common/Utilities/ActionResponse.cs
@@ -3,7 +3,6 @@ using ChustaSoft.Common.Enums;
 using System.Collections.Generic;
 using System.Runtime.Serialization;
 
-
 namespace ChustaSoft.Common.Utilities
 {
