+++ ChustaSoft.Common/Utilities/PaginatedList.cs
@@ -4,7 +4,6 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Runtime.Serialization;
 
-
 namespace ChustaSoft.Common.Utilities
 {
