+++ ChustaSoft.Common/Utilities/ThreadSafeList.cs
@@ -1,7 +1,6 @@
 ﻿using System.Collections;
 using System.Collections.Generic;
 
-
 namespace ChustaSoft.Common.Utilities
 {
